requirements.txt
django==1.4.1

mock==1.0b1
factory_boy==1.2.0

south==0.7.6
django-autoslug==1.5.0
unidecode
django-jsonfield==0.8.11

django-compressor==1.2
slimit==0.7.4

celery==3.0.6
django-celery==3.0.6

requests==0.13.6
requests-oauth==0.4.1

# Gondor requirements
gondor==1.1.4
psycopg2==2.4.5
gunicorn==0.14.3
